requirements.txt
Django==6.0.2
Pillow==10.1.0
python-decouple==3.8
requests

